Moving InclinedExpoenential class out of base and into its own module
