+++ galsim/base.py
@@ -2896,113 +2896,6 @@ _galsim.SBSpergel.__repr__ = lambda self: \
         'galsim._galsim.SBSpergel(%r, %r, %r, %r, %r)' % self.__getinitargs__()
 
 
-class InclinedExponential(GSObject):
-    """A class describing an inclined exponential profile.
-
-    The Inclined Exponential surface brightness profile is characterized by three properties: its
-    inclination angle 'i' (where 0 = face-on and pi/2 = edge-on), its scale radius, and its scale
-    height. The 3D light distribution function is:
-
-    I(R,z) = I_0 / (2h_s) * sech^2 (z/h_s) * exp(-R/R_s)
-
-    where z is the distance along the minor axis, R is the radial distance from the minor axis,
-    R_s is the scale radius of the disk, h_s is the scale height of the disk, and I_0 is the central
-    surface brightness of the face-on disk.
-
-    In this implementation, the profile is inclined along the y-axis. This means that it will likely
-    need to be rotated in most circumstances.
-
-    At present, this profile is not enabled for photon-shooting.
-
-    Initialization
-    --------------
-
-    @param inclination      The inclination angle, i
-    @param scale_radius     The scale radius of the exponential disk.  Typically given in arcsec.
-    @param scale_height     The scale height of the exponential disk.  Typically given in arcsec.
-    @param flux             The flux (in photons) of the profile. [default: 1]
-    @param gsparams         An optional GSParams argument.  See the docstring for GSParams for
-                            details. [default: None]
-
-    Methods
-    -------
-
-    In addition to the usual GSObject methods, InclinedExponential has the following access methods:
-
-        >>> inclination = inclined_exponential_obj.getInclination()
-        >>> r0 = inclined_exponential_obj.getScaleRadius()
-        >>> h0 = inclined_exponential_obj.getScaleHeight()
-    """
-    _req_params = { "inclination" : float, "scale_radius" : float, "scale_height" : float }
-    _opt_params = { "flux" : float }
-    _takes_rng = False
-
-    def __init__(self, inclination, scale_radius, scale_height,
-                 flux=1., gsparams=None):
-
-        if not isinstance(inclination, galsim.Angle):
-            raise TypeError("Input inclination should be an Angle")
-
-        i_rad = inclination.rad()
-
-        GSObject.__init__(self, _galsim.SBInclinedExponential(i_rad, scale_radius, scale_height, flux,
-                                                              gsparams))
-        self._gsparams = gsparams
-
-    def getInclination(self):
-        """Return the inclination angle for this profile in radians.
-        """
-        return self.SBProfile.getInclination()
-
-    def getScaleRadius(self):
-        """Return the scale radius for this Sersic profile.
-        """
-        return self.SBProfile.getScaleRadius()
-
-    def getScaleHeight(self):
-        """Return the scale radius for this Sersic profile.
-        """
-        return self.SBProfile.getScaleHeight()
-
-    @property
-    def inclination(self): return self.getInclination()
-    @property
-    def scale_radius(self): return self.getScaleRadius()
-    @property
-    def scale_height(self): return self.getScaleHeight()
-
-    def __eq__(self, other):
-        return ((isinstance(other, galsim.InclinedExponential) and
-                 (self.inclination == other.inclination) and
-                 (self.scale_radius == other.scale_radius) and
-                 (self.scale_height == other.scale_height) and
-                 (self.flux == other.flux) and
-                 (self._gsparams == other._gsparams)))
-
-    def __hash__(self):
-        return hash(("galsim.InclinedExponential", self.inclination, self.scale_radius, self.scale_height, self.flux,
-                     self._gsparams))
-
-    def __repr__(self):
-        return 'galsim.InclinedExponential(i=%r, scale_radius=%r, scale_height=%r, flux=%r, gsparams=%r)' % (
-            self.inclination, self.scale_radius, self.scale_height, self.flux, self._gsparams)
-
-    def __str__(self):
-        s = 'galsim.InclinedExponential(i=%s, scale_radius=%s, scale_height=%s' % (self.inclination, self.scale_radius,
-                                                                                 self.scale_height)
-        if self.flux != 1.0:
-            s += ', flux=%s' % self.flux
-        s += ')'
-        return s
-
-_galsim.SBInclinedExponential.__getinitargs__ = lambda self: (
-        self.getInclination(), self.getScaleRadius(), self.getScaleHeight(), self.getFlux(), self.getGSParams())
-_galsim.SBInclinedExponential.__getstate__ = lambda self: None
-_galsim.SBInclinedExponential.__setstate__ = lambda self, state: 1
-_galsim.SBInclinedExponential.__repr__ = lambda self: \
-        'galsim._galsim.SBInclinedExponential(%r, %r, %r, %r, %r)' % self.__getinitargs__()
-
-
 # Set the docstring for GSParams here.  It's easier to edit in the python layer than using
 # the boost python doc parameter.
 GSParams.__doc__ = """
